Allow docs to be used as a python module
In an effort to split documentation off the main repository
documentation can also be viewed as a Python package. It would probably
make better sense to bundle docs with ArC1 however the documentation
depends on pandoc and weasyprint which admittedly can be a bit of a
hassle on Windows. To alleviate this documentation can be distributed as
a "binary" wheel and used as "extras_require" in ArC1.